servers/azure-resource-graph-mcp/requirements.txt
mcp>=1.3.0
azure-identity>=1.17.0
azure-mgmt-resourcegraph>=8.0.0
azure-mgmt-subscription>=3.1.1
